python_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
openai==1.44.0
langchain==0.2.14
langchain-openai==0.1.23
langchain-core==0.2.38
langchain-community==0.2.12